+++ src/Classes/FinalRule.php
@@ -120,7 +120,13 @@ final class FinalRule implements Rule
             return false;
         }
 
-        if (\is_int(\preg_match_all('/@(\S+)(?=\s|$)/', $docComment->getReformattedText(), $matches))) {
+        $reformattedComment = $docComment->getReformattedText();
+
+        if (!\is_string($reformattedComment)) {
+            return false;
+        }
+
+        if (\is_int(\preg_match_all('/@(\S+)(?=\s|$)/', $reformattedComment, $matches))) {
             foreach ($matches[1] as $annotation) {
                 foreach (self::$whitelistedAnnotations as $whitelistedAnnotation) {
                     if (0 === \mb_strpos($annotation, $whitelistedAnnotation)) {