+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-For a full diff see [`0.15.3...main`][0.15.3...main].
+For a full diff see [`1.0.0...main`][1.0.0...main].
+
+## [`1.0.0`][1.0.0]
+
+For a full diff see [`0.15.3...1.0.0`][0.15.3...1.0.0].
 
 ### Changed
 
@@ -358,6 +362,7 @@ For a full diff see [`362c7ea...0.1.0`][362c7ea...0.1.0].
 [0.15.1]: https://github.com/ergebnis/phpstan-rules/releases/tag/0.15.1
 [0.15.2]: https://github.com/ergebnis/phpstan-rules/releases/tag/0.15.2
 [0.15.3]: https://github.com/ergebnis/phpstan-rules/releases/tag/0.15.3
+[1.0.0]: https://github.com/ergebnis/phpstan-rules/releases/tag/1.0.0
 
 [362c7ea...0.1.0]: https://github.com/ergebnis/phpstan-rules/compare/362c7ea...0.1.0
 [0.1.0...0.2.0]: https://github.com/ergebnis/phpstan-rules/compare/0.1.0...0.2.0
@@ -386,7 +391,8 @@ For a full diff see [`362c7ea...0.1.0`][362c7ea...0.1.0].
 [0.15.0...0.15.1]: https://github.com/ergebnis/phpstan-rules/compare/0.15.0...0.15.1
 [0.15.1...0.15.2]: https://github.com/ergebnis/phpstan-rules/compare/0.15.1...0.15.2
 [0.15.2...0.15.3]: https://github.com/ergebnis/phpstan-rules/compare/0.15.2...0.15.3
-[0.15.3...main]: https://github.com/ergebnis/phpstan-rules/compare/0.15.3...main
+[0.15.3...1.0.0]: https://github.com/ergebnis/phpstan-rules/compare/0.15.3...1.0.0
+[1.0.0...main]: https://github.com/ergebnis/phpstan-rules/compare/1.0.0...main
 
 [#1]: https://github.com/ergebnis/phpstan-rules/pull/1
 [#4]: https://github.com/ergebnis/phpstan-rules/pull/4